+++ packages/rest-hooks/src/restClient.ts
@@ -1,4 +1,4 @@
-import axios, { AxiosResponse, InternalAxiosRequestConfig } from 'axios'
+import axios, { AxiosError, AxiosResponse, InternalAxiosRequestConfig } from 'axios'
 import { QueryFunctionContext } from '@tanstack/react-query'
 
 import { assocPath, keys, toUpper } from 'ramda'
@@ -69,7 +69,6 @@ export const createRestClient: StcRest.CreateRestClient = (
   }
 
   // all clients use these middlewares
-
   restClient.axiosClient.interceptors.request.use(
     _requestPreprocessor({
       verbose: !!clientConfig?.verbose,
@@ -78,10 +77,19 @@ export const createRestClient: StcRest.CreateRestClient = (
   )
 
   restClient.axiosClient.interceptors.response.use(
-    _responsePostProcessor({
-      verbose: !!clientConfig?.verbose,
-      responsePostProcessor: clientConfig?.responsePostProcessorFn
-    })
+    (response) => {
+      // onSuccess response handling
+      return onPostProcessorResponseSuccess({
+        verbose: !!clientConfig?.verbose,
+        responsePostProcessor: clientConfig?.responsePostProcessorFn,
+      })(response)
+    },
+    // OnError response handling
+    (error) => {
+      return onPostProcessorResponseError({
+        error, onAuthFailureFn: clientConfig?.onAuthFailureFn
+      })(error)
+    }
   )
   return restClient
 }
@@ -115,9 +123,14 @@ const _requestPreprocessor =
         : req
     }
 
-interface _RequestPostProcessorOptions {
+interface _RequestPostProcessorOptionsSuccess {
   verbose?: boolean
   responsePostProcessor?: (rsp: AxiosResponse) => AxiosResponse
+}
+
+interface _RequestPostProcessorOptionsError {
+  error: any
+  onAuthFailureFn?: (error: any) => void
 }
 
 export const f = async () => {
@@ -126,8 +139,8 @@ export const f = async () => {
 
 
 
-const _responsePostProcessor =
-  (opts: _RequestPostProcessorOptions) =>
+const onPostProcessorResponseSuccess =
+  (opts: _RequestPostProcessorOptionsSuccess) =>
     (rsp: AxiosResponse) =>
     {
 
@@ -141,6 +154,17 @@ const _responsePostProcessor =
 
       return responsePostProcessor(rsp)
     }
+
+
+const onPostProcessorResponseError =
+   (opts: _RequestPostProcessorOptionsError) =>
+     (rspError: AxiosError) => {
+       const { error, onAuthFailureFn } = opts
+       if (error.status === 401 && onAuthFailureFn) {
+         onAuthFailureFn(error)
+       }
+       return rspError
+     }
 
 /**
  * @function _expandRestPath